Learning_process_level/Task_self_transformation/DeCap/requirements.txt
diffusers==0.29.0
ftfy==6.3.1
numpy==1.26.4
packaging==25.0
Pillow==11.3.0
regex==2024.11.6
scikit_opt==0.6.6
tqdm==4.67.0
transformers==4.47.0
